move shemas to a mixin
